update change password

+++ resources/views/dashboard/siswa.blade.php
@@ -25,10 +25,10 @@
                     </div>
             </div>
             <div class="card-body">
-                <table class="table table-sm table-bordered table-hover ">
+                <table class="table table-sm table-bordered table-hover table-responsive">
                     <thead>
                       <tr>
-                        <th scope="col">No</th>
+                        {{-- <th scope="col">No</th> --}}
                         <th scope="col">NISN</th>
                         <th scope="col">Nama Siswa</th>
                         <th scope="col">Kelas</th>
@@ -38,17 +38,17 @@
                     <tbody id="data-siswa">
                         @if ($siswa->count() > 0)
                         @foreach ($siswa as $s)
-                            
                         <tr id="siswa_id_{{$s->id}}">
-                            <th scope="row">{{$s->id}}</th>
+                            {{-- <th scope="row">{{$s->id}}</th> --}}
                             <td>{{$s->nis}}</td>
                             <td>{{$s->name}}</td>
                             <td>{{$s->kelas}}</td>
                             <td class="text-center" width="1%">
                                 
                                     <div class="btn-group mr-2" role="group" aria-label="First group">
-                                        <button class="btn btn-danger btn-sm" id="delete-data" data-id="{{$s->id}}"><i class="fa fa-trash"></i></button>
-                                        <button class="btn btn-primary btn-sm" id="edit-data" data-id="{{$s->id}}"><i class="fa fa-pen"></i></button>
+                                        <button class="btn btn-primary btn-sm edit-data" id="edit-data" data-id="{{$s->id}}" title="Edit Siswa"><i class="fa fa-pen"></i></button>
+                                        <button class="btn btn-info btn-sm edit-password" id="edit-password" data-id="{{$s->id}}" title="Ganti Password"><i class="fa fa-lock"></i></button>
+                                        <button class="btn btn-danger btn-sm delete-data" id="delete-data" data-id="{{$s->id}}" title="Hapus Siswa"><i class="fa fa-trash"></i></button>
                                     </div>
                                 
                             </td>
@@ -154,6 +154,43 @@
       </div>
     </div>
   </div>
+  {{-- end modal --}}
+
+  {{-- modal ganti password --}}
+  <div class="modal fade" id="modalPass" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
+    <div class="modal-dialog" role="document">
+      <div class="modal-content">
+        <div class="modal-header">
+          <h5 class="modal-title"></h5>
+          <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+            <span aria-hidden="true">&times;</span>
+          </button>
+        </div>
+        <div class="modal-body">
+          <form id="formPass" name="formPass">
+              <input type="hidden" name="id_siswa" id="idUser" value="">
+              <div class="row">
+
+                <div class="col-md-12">
+                  <div class="input-group">
+                        <input type="password" class="form-control" id="password" name="password" required placeholder="Masukkan Password Baru">
+                        <div class="input-group-append">    
+                          <span class="input-group-text"><i class="fa fa-lock"></i></span>
+                        </div>
+                  </div>
+                  
+                </div>               
+          </div>
+        </div>
+        <div class="modal-footer">
+          <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+          <button type="sumbit" class="btn btn-primary" id="buttonPass"></button>
+        </div>
+        </form>
+      </div>
+    </div>
+  </div>
+  {{-- end modal --}}
 
 <script>
     $(document).ready(function(){
@@ -169,7 +206,7 @@
             $("#action-button").text("Tambah Data");
         });
 
-        $("#edit-data").click(function(){
+        $(".edit-data").click(function(){
             var siswa_id=$(this).data('id');
             $.get('siswa/'+siswa_id+'/edit',function(data){
             $("#modalaction").modal("show");
@@ -177,26 +214,36 @@
             $("#action-button").text("Simpan Perubahan");
             $("#name").val(data.name);
             $("#nis").val(data.nis);
-
+            $("#kelas").val(data.id_kelas);
             });
         })
 
-        $("#delete-data").click(function(){
+        $(".delete-data").click(function(){
             var id_siswa=$(this).data('id');
-            confirm ("Apakah Anda Yakin?");
+            var cek = confirm ("Apakah Anda Yakin?");
+            if (cek == true) {
+              $.ajax({
+              type: "DELETE",
+              url: "siswa/"+id_siswa,
+              success: function (data) {
+                  $("#siswa_id_" + id_siswa).remove();
+                },
+                error: function (data) {
+                    console.log('Error:', data);
+                }
+              });
+            } 
+            
 
-            $.ajax({
-            type: "DELETE",
-            url: "siswa/"+id_siswa,
-            success: function (data) {
-                $("#siswa_id_" + id_siswa).remove();
-            },
-            error: function (data) {
-                console.log('Error:', data);
-            }
-        });
-
-        });
+          });
+          $(".edit-password").click(function(){
+            var siswa_id=$(this).data('id');
+            // alert(siswa_id);
+            $("#modalPass").modal("show");
+            $(".modal-title").text("Ganti Password");
+            $("#buttonPass").text("Simpan");
+            $("#idUser").val(siswa_id);
+          })
     });
 //     if ($("#formsiswa").length > 0) {
 //       $("#formsiswa").validate({
@@ -244,10 +291,51 @@ if ($("#formsiswa").length > 0) {
       $('#action-button').html('Sending..');
 
       data= $('#formsiswa').serialize();
-      alert(data);
+      // alert(data);
       $.ajax({
           data: $('#formsiswa').serialize(),
           url: "{{ route('siswa.store') }}",
+          type: "POST",
+          dataType: 'json',
+          success: function (data) {
+            // alert(data);
+            console.log(data);
+            // var siswa='<tr id="siswa_id_'+data.id+'"><th scope="row">'+ data.id +'</th><td>'+ data.nis +'</td><td>'+ data.name +'</td><td>'+ data.kelas +'</td>';
+            //  siswa+='<td class="text-center" width="1%"><div class="btn-group mr-2" role="group" aria-label="First group"><button class="btn btn-danger btn-sm" id="delete-data" data-id="'+ data.id +'"><i class="fa fa-trash"></i></button><button class="btn btn-primary btn-sm" id="edit-data" data-id="'+ data.id +'"><i class="fa fa-pen"></i></button></div></td></tr>';
+               
+              
+            //   if (actionType == "tambah-data") {
+            //       $('#siswa_id').prepend(post);
+            //   } else {
+            //       $("#siswa_id_" + data.id).replaceWith(post);
+            //   }
+              $('#formsiswa').trigger("reset");
+              $('#modalaction').modal('hide');
+              $("#action-button").text("Tambah Data");
+              location.reload();
+          },
+          error: function (data) {
+              // alert(data->responseText);
+              console.log('Error:', data);
+              $('#btn-save').html('Save Changes');
+          }
+      });
+    }
+  })
+}
+if ($("#formPass").length > 0) {
+      $("#formPass").validate({
+ 
+     submitHandler: function(form) {
+
+      var actionType = $('#buttonPass').val();
+      $('#buttonPass').html('Sending..');
+
+      data= $('#formPass').serialize();
+      alert(data);
+      $.ajax({
+          data: $('#formPass').serialize(),
+          url: "{{ route('siswa.gantipass') }}",
           type: "POST",
           dataType: 'json',
           success: function (data) {
